Algebra.Algorithms.Groebner.SelectionStrategy

Documentation

class SelectionStrategy n s where

Type-class for selection strategies in Buchberger's algorithm.

Associated Types

type Weight n s ord :: Type

Methods

calcWeight :: (IsOrderedPolynomial poly, n ~ Arity poly) => Proxy s -> poly -> poly -> Weight n s (MOrder poly)

Calculates the weight for the given pair of polynomial used for selection strategy.

calcWeight' :: (SelectionStrategy (Arity poly) s, IsOrderedPolynomial poly) => s -> poly -> poly -> Weight (Arity poly) s (MOrder poly)

Calculate the weight of given polynomials w.r.t. the given strategy. Buchberger's algorithm proccesses the pair with the most least weight first. This function requires the Ord instance for the weight; this constraint is not required in the calcWeight because of the ease of implementation. So use this function.

data GrevlexStrategy

Choose the pair with the least LCM(LT(f), LT(g)) w.r.t. Grevlex order.

Constructors

GrevlexStrategy

data NormalStrategy

Buchberger's normal selection strategy. This selects the pair with the least LCM(LT(f), LT(g)) w.r.t. current monomial ordering.

Constructors

NormalStrategy

newtype SugarStrategy s

Sugar strategy. This chooses the pair with the least phantom homogenized degree and then break the tie with the given strategy (say s).

Constructors

SugarStrategy s

data GradedStrategy

Constructors

GradedStrategy