Algebra.Ring.Polynomial.Quotient

Documentation

data Quotient poly ideal

The polynomial modulo the ideal indexed at the last type-parameter.

data QIdeal poly

reifyQuotient :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => Ideal poly -> (forall (ideal :: Type). Reifies ideal (QIdeal poly) => Proxy ideal -> a) -> a

Reifies the ideal at the type-level. The ideal can be recovered with reflect.

modIdeal :: forall poly ideal. (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => poly -> Quotient poly ideal

Polynomial modulo ideal.

modIdeal' :: (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => Proxy ideal -> poly -> Quotient poly ideal

Polynomial modulo ideal given by Proxy.

quotRepr :: Quotient poly ideal -> poly

Representative polynomial of given quotient polynomial.

withQuotient :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => Ideal poly -> (forall (ideal :: Type). Reifies ideal (QIdeal poly) => Quotient poly ideal) -> poly

Computes polynomial modulo ideal.

vectorRep :: forall poly ideal. (IsOrderedPolynomial poly, Reifies ideal (QIdeal poly)) => Quotient poly ideal -> Vector (Coefficient poly)

genQuotVars :: forall poly ideal. (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => [Quotient poly ideal]

genQuotVars' :: forall poly ideal. (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => Proxy ideal -> [Quotient poly ideal]

gBasis' :: Reifies ideal (QIdeal poly) => Proxy ideal -> [poly]

matRep0 :: forall poly ideal. (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => Proxy poly -> Proxy ideal -> OrderedMonomial (MOrder poly) (Arity poly) -> Matrix (Coefficient poly)

standardMonomials :: forall poly ideal. (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => Maybe [Quotient poly ideal]

standardMonomials' :: (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => Proxy ideal -> Maybe [Quotient poly ideal]

Find the standard monomials of the quotient ring for the zero-dimensional ideal, which are form the basis of it as k-vector space.

matRepr' :: forall poly ideal. (Field (Coefficient poly), Reifies ideal (QIdeal poly), IsOrderedPolynomial poly) => Quotient poly ideal -> Matrix (Coefficient poly)

reduce :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => poly -> Ideal poly -> poly

Reduce polynomial modulo ideal.

multWithTable :: (IsOrderedPolynomial poly, Reifies ideal (QIdeal poly)) => Quotient poly ideal -> Quotient poly ideal -> Quotient poly ideal

multUnamb :: (IsOrderedPolynomial poly, Field (Coefficient poly), Reifies ideal (QIdeal poly)) => Quotient poly ideal -> Quotient poly ideal -> Quotient poly ideal

isZeroDimensional :: (IsOrderedPolynomial poly, Field (Coefficient poly)) => [poly] -> Bool