Algebra.Ring.Polynomial.Labeled

Documentation

class UniqueList xs => IsUniqueList (xs :: [Symbol])

data LabPolynomial poly (vars :: [Symbol]) where

Bundled Patterns

pattern LabelPolynomial :: Wraps vars poly => Wraps vars poly => poly -> LabPolynomial poly vars

type family LabPolynomial' r ord vars where ...

Convenient type-synonym for LabPlynomial wrapping OrderedPolynomial and Unipol.

Equations

LabPolynomial' r ord '[x] = LabPolynomial (Unipol r) '[x]

LabPolynomial' r ord vars = LabPolynomial (OrderedPolynomial r ord (Length vars)) vars

type LabUnipol r sym = LabPolynomial (Unipol r) '[sym]

Convenient type-synonym for LabPlynomial wrapping univariate polynomial Unipol.

type Wraps vars poly = (IsUniqueList vars, Arity poly ~ Length vars)

canonicalMap :: forall xs ys poly poly'. (SingI xs, SingI ys, IsSubsetOf xs ys, Wraps xs poly, Wraps ys poly', IsPolynomial poly, IsPolynomial poly', Coefficient poly ~ Coefficient poly') => LabPolynomial poly xs -> LabPolynomial poly' ys

canonicalMap' :: (SingI xs, SingI ys, IsSubsetOf xs ys, Wraps xs poly, Wraps ys poly', IsPolynomial poly, IsPolynomial poly', Coefficient poly ~ Coefficient poly') => proxy poly' -> proxy' ys -> LabPolynomial poly xs -> LabPolynomial poly' ys

renameVars :: (SingI ys, SingI xs, Length xs ~ Length ys, IsPolynomial poly) => LabPolynomial poly xs -> LabPolynomial poly ys

renameVars' :: (SingI ys, SingI xs, Length xs ~ Length ys, IsPolynomial poly) => proxy ys -> LabPolynomial poly xs -> LabPolynomial poly ys

class All ((FlipSym0 @@ ElemSym0) @@ ys) xs ~ 'True => IsSubsetOf (xs :: [a]) (ys :: [a])